// src/api-client/models/surplus-package-sort.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const SurplusPackageSort = {
    Distance: 'distance',
    Urgency: 'urgency',
    Weight: 'weight',
    Value: 'value'
} as const;

export type SurplusPackageSort = typeof SurplusPackageSort[keyof typeof SurplusPackageSort];